minor change to rename test.utils to test.helpers to avoid confusion
