+++ test/catalog/test_semistruct_csv.py
@@ -1,5 +1,5 @@
 import unittest
-from test.utils import CatalogHelper, BaseTestCase
+from test.helpers import CatalogHelper, BaseTestCase
 
 
 class TestSemistructuredCsv (BaseTestCase):
