+++ test/helpers.py
@@ -1,3 +1,5 @@
+"""Helpers for the tests."""
+
 import abc
 import csv
 import logging
