+++ test/optimizer/test_consolidate.py
@@ -1,4 +1,4 @@
-from test.utils import CatalogHelper, BaseTestCase
+from test.helpers import CatalogHelper, BaseTestCase
 
 
 class TestConsolidate (BaseTestCase):
