+++ test/util/__init__.py
@@ -1,0 +1,4 @@
+"""A package of tests of the `chisel.util` package.
+
+Not to be confused with utilities for the chisel tests. For those, see the `test.helpers` module.
+"""
